chore: get fee asset refactor (#4198)
