+++ src/state/slices/assetsSlice/selectors.ts
@@ -1,20 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit'
 import type { Asset } from '@shapeshiftoss/asset-service'
-import type {
-  AssetId,
-  AssetReference,
-  ChainId,
-  ChainNamespace,
-  ChainReference,
-} from '@shapeshiftoss/caip'
-import {
-  ASSET_REFERENCE,
-  CHAIN_NAMESPACE,
-  CHAIN_REFERENCE,
-  fromAssetId,
-  fromChainId,
-  toAssetId,
-} from '@shapeshiftoss/caip'
+import type { AssetId, ChainId } from '@shapeshiftoss/caip'
+import { fromAssetId } from '@shapeshiftoss/caip'
 import cloneDeep from 'lodash/cloneDeep'
 import sortBy from 'lodash/sortBy'
 import createCachedSelector from 're-reselect'
@@ -25,6 +12,7 @@ import { selectAssetIdParamFromFilter } from 'state/selectors'
 import { selectCryptoMarketDataIdsSortedByMarketCap } from 'state/slices/marketDataSlice/selectors'
 
 import { assetIdToFeeAssetId } from '../portfolioSlice/utils'
+import { getFeeAssetByAssetId, getFeeAssetByChainId } from './utils'
 
 export const selectAssetById = createCachedSelector(
   (state: ReduxState) => state.assets.byId,
@@ -81,82 +69,14 @@ export const selectChainIdsByMarketCap = createDeepEqualOutputSelector(
     }, []),
 )
 
-// @TODO figure out a better way to do this mapping. This is a stop gap to make selectFeeAssetById
-const chainIdFeeAssetReferenceMap = (
-  chainNamespace: ChainNamespace,
-  chainReference: ChainReference,
-): AssetReference => {
-  return (() => {
-    switch (chainNamespace) {
-      case CHAIN_NAMESPACE.Utxo:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.BitcoinMainnet:
-            return ASSET_REFERENCE.Bitcoin
-          case CHAIN_REFERENCE.BitcoinCashMainnet:
-            return ASSET_REFERENCE.BitcoinCash
-          case CHAIN_REFERENCE.DogecoinMainnet:
-            return ASSET_REFERENCE.Dogecoin
-          case CHAIN_REFERENCE.LitecoinMainnet:
-            return ASSET_REFERENCE.Litecoin
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      case CHAIN_NAMESPACE.Evm:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.AvalancheCChain:
-            return ASSET_REFERENCE.AvalancheC
-          case CHAIN_REFERENCE.EthereumMainnet:
-            return ASSET_REFERENCE.Ethereum
-          case CHAIN_REFERENCE.OptimismMainnet:
-            return ASSET_REFERENCE.Optimism
-          case CHAIN_REFERENCE.BnbSmartChainMainnet:
-            return ASSET_REFERENCE.BnbSmartChain
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      case CHAIN_NAMESPACE.CosmosSdk:
-        switch (chainReference) {
-          case CHAIN_REFERENCE.CosmosHubMainnet:
-            return ASSET_REFERENCE.Cosmos
-          case CHAIN_REFERENCE.OsmosisMainnet:
-            return ASSET_REFERENCE.Osmosis
-          case CHAIN_REFERENCE.ThorchainMainnet:
-            return ASSET_REFERENCE.Thorchain
-          default:
-            throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-        }
-      default:
-        throw new Error(`Chain namespace ${chainNamespace} on ${chainReference} not supported.`)
-    }
-  })()
-}
-
 export const selectFeeAssetByChainId = createCachedSelector(
   selectAssets,
   (_state: ReduxState, chainId: ChainId) => chainId,
-  (assetsById, chainId): Asset | undefined => {
-    if (!chainId) return undefined
-    const { chainNamespace, chainReference } = fromChainId(chainId)
-    const feeAssetId = toAssetId({
-      chainId,
-      assetNamespace: 'slip44',
-      assetReference: chainIdFeeAssetReferenceMap(chainNamespace, chainReference),
-    })
-    return assetsById[feeAssetId]
-  },
+  (assetsById, chainId): Asset | undefined => getFeeAssetByChainId(assetsById, chainId),
 )((_state: ReduxState, chainId) => chainId ?? 'chainId')
 
 export const selectFeeAssetById = createCachedSelector(
   selectAssets,
   (_state: ReduxState, assetId: AssetId) => assetId,
-  (assetsById, assetId): Asset | undefined => {
-    const { chainNamespace, chainReference } = fromAssetId(assetId)
-    const feeAssetId = toAssetId({
-      chainNamespace,
-      chainReference,
-      assetNamespace: 'slip44',
-      assetReference: chainIdFeeAssetReferenceMap(chainNamespace, chainReference),
-    })
-    return assetsById[feeAssetId]
-  },
+  (assetsById, assetId): Asset | undefined => getFeeAssetByAssetId(assetsById, assetId),
 )((_s: ReduxState, assetId: AssetId) => assetId ?? 'assetId')
